Backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
Flask>=2.0
numpy>=1.22
pillow>=8.0.0
tensorflow>=2.10
huggingface-hub>=0.14.1
gunicorn>=20.0.0
h5py>=3.7.0
werkzeug>=2.0
Pillow==10.0.1
numpy==1.24.3






